Initial (and late) commit
The basics of my Sudoku class is written, with read and write access to
the squares and storage for the domains of each square. There is also a
main file with some minor testing, and a new tests directory that I
created to begin trying unit testing using Boost. I've also
incorporated CMake into the project to simplify building it