Find all solutions, add helper functions, add dimension + checks
Added a function to find all solutions to a sudoku if there are
multiple. Added some more helper functions and added checks to see if
read and writes are within the dimensions of the sudoku puzzle.

Also added the dimension variable to Sudoku so that puzzles may be
variable size.